// Projects/Full Stack Application/Frontend/src/app/models/customer.model.ts
export interface Customer {
    customerId: number; 
    firstName: string;
    middleName: string;
    lastName: string;
    dob: Date;
    addressLine1: string;
    addressLine2: string;
    city: string;
    state: string;
    zipCode: number;
    country: string;
    email: string;
    primaryContactNo: number;
    secondaryContactNo: number;
    gender: string;
    department: string;
    maritalStatus: string;
    bankName: string;
    bankBranch: string;
    qualification: string;
    termsAccepted: boolean;
  }
  